Add files via upload
Near final version upload. Responsiveness improved and completed on all associated web pages. Structure of internal files fixed. Waivers and Location information were created and linked to their respective buttons on the site. Iteration 1 is included in the repo and debug and test builds are included in a compressed folder.